Delete mdp_dist. All future updates should be made in release-manager branch.
From now on mdp_dist will only exist in the release-manager branch.
If the release manager decides to update, the specific commit of the master branch
should be merged to the release-manager branch. (Of course, care has to be taken
that the mdp_dist folder is not removed from release-manager after merging.)