Introducing a node for a simple discrete Hopfield model.
The node is currently implemented as a classifier because that is what
it actually does. However, because feature space is a superset of label
space and the classification does not give any probabilities, it is also
possible to consider this not a classifier.

On the other hand, the output of the node is not deterministic and
classification results may vary in different runs. Therefore, the
classification output might not be sensible to occur in an unsupervised
flow. To emphasise this intention, the node is made a classifier.