ERF: simplify testing by providing commonly used imports in test/_tools
All testing files can just say
  from _tools import *
and the get numx, numx_rand, mult, ...

Py.test is not imported into _tools, so that it is easier to possible
to import files which don't actually depend on py.test when it is not
available.

Also includes some tiny naming and space corrections.