Fix --seed option for py.test >= 1.3
Py.test in version 1.3 added a feature to skip conftest.py files when
they are exactly the same. This helps when a directory is copied to a
subdirectory of itself and the config files are not idempotent.
Unfortunately, this runs afoul of our logic to add the --seed option
everywhere, irrespective of wheter py.test is run implicitely for the
whole mdp-toolkit or for mdp or for bimdp or for both. The solution is
to modify one of the conftest.py files to be unidentical.

Tested with with py.test 1.2.1 and 1.3.3.