FIX: Patch for using deepcopy instead of pickle for Node and Flow
Hi all,

here is a patch and test script for replacing the pickle with deepcopy
in Node.copy() and Flow.copy(). The signature of the copy function
changes (the protocol keyword is no longer necessary), I don't know
what the proper way to handle this is (deprecation error?), so I'll
leave it up to you guys.

Thanks for all the efforts, we get very positive feedback from all
users of mdp and Oger!

David