FIX: test_seed failured when testing running bimdp.test() after mdp.test()
The reason was that when the test functions were imported from
mdp.test into bimdp.test, they still referenced mdp.test._SEED
global variable. But the --seed option in bimdp is independent,
so it was set to a different random seed. By duplicating
test_seed.py, a second independent global is created.