FIX: clean up __all__ in mdp/__init__.py
numx* should not be in all. Quoting Tiziano:

> numx was not in __all__, and therefore was not documented by epydoc.
> Once it is added to __all__, indeed things break heavily.
well, I think even it was by mistake, it is a good thing that numx
is not in __all__. mdp.numx is an *internal* thing. people doing
from mdp import * should not get numx. so, leave it like this and we
don't even need a workaround for epydoc.

While at it, remove whitening which was already gone and make the list
alphabetical.