FIX: fixed a bit pp_remote support.
I commented out the test. There's no point in having a test that
requires a working ssh setup. we should move that code in the examples.
TODO: I still did not manage to test the remote support properly:
it does not work if ssh spits some session information on startup :-(