FIX: fix wrong normalization in SFANode
SFANode was using the covariance matrix of the derivatives,
while the SFA algorithm requires the non-centered second moment
matrix of the derivatives. This bug is probably irrelevant in most
cases, but in some corner cases the error scales with 1/sqrt(N) where
N is the number of samples. Many thanks to Sven Dähne for fishing this
out ;-)