ENH: revert the condition for pp monkeypatching
It's better to only enable it when explicitely requested.
The warning is also removed, because it doesn't make sense
to warn about something that was explicitely requested.
If the option to monkey patch is not given, check if pp
is likely to be broken, and if so, disable it at configuration
time.