Add mdp.__revision__ to mdp.config.info() output
This changes the order of initialization a bit: first
mdp.configuration is fully imported, then mdp.utils. Because of this change,
mdp.utils.repo_revision was renamed to mdp.repo_revision. It
is imported only in mdp.configuration and then deleted from the visible
namespace. get_git_revision was available as utils.get_git_revision,
and now it's gone, but mdp.__revision__ contains the same value, so there
little need to export the function. I hope nobody used it.

>>> print mdp.config.info()
          python: 2.7.2.final.0
             mdp: MDP-3.1-74-gbaca2a8+
 parallel python: 1.6.0
          shogun: NOT AVAILABLE: No module named shogun
          libsvm: libsvm.so.3
          joblib: 0.4.6
         scikits: 0.8.1
            numx: scipy 0.9.0
          symeig: scipy.linalg.eigh