FIX: no with statement in tests {flow,node}.save
Apparently windows does not allow one process to open a file multiple
times. As {flow,node}.save use internally the with statement to open
the dump file, we can not use the with statement in the test of these
methods: nesting with statement means opening the same file multiple
times.