FIX: use two identical conftest.py for mdp and bimdp.
We can not ship only one conftest.py for mdp and bimdp, because when
you install using setup.py the file conftest.py is not going to be installed
[you only install files below the top directories mdp and bimdp].
In order for py.test to use the same configuration and have the same options
for both mdp and bimdp, we need to have two identical copies of the file
in mdp/test/conftest.py and bimdp/test/conftest.py. We can not use symlinks
because they do not work in git under windows.