Revert "Revert back to original triple conftest.py setup"
Revert back to use only two conftest.py file. The reasoning
being that the added value of the three file configuration is not worth
the higher complexity of the code that is needed to support it.

The main advantage of having the configuration that we revert here is
that you can run py.test --seed N (without additional arguments) in
the MDP top-level directory. With the configuration introduced by the
present commit, you will need to run "py.test --seed N mdp" or
"py.test --seed N mdp bimdp" [this is to force py.test to load the
mdp/test/conftest.py where the seed option has been defined].
Note that if you do not need the seed option, you can stil run pu.test
without arguments.

For the end user there's no difference, as they should keep on using:

>>> import mdp
>>> mdp.test(seed=N)
>>> import bimdp
>>> bimdp.test(seed=N)

which has always worked.

It seems that this issue only affects and bothers Zbyszek and me, so the
rest of the world can ignore the whole thing ;)

This reverts commit e2c180055bea33a0114cac462dbc460e999d2376.