FIX: add two theoretically useless encoding marks
... in hope that they help epydoc.

After the __module__ attributes are properly faked, epydoc and sphinx
actually look at the exporting module for the information encoding.
When utf-8 encoding marks are added, epydoc and sphinx think that all
exported classes' docstrings are unicode. This is OK, because
everything will be either be either ASCII or UTF-8.