ENH: recurse into class attributes when doing namespace fixups
After this change, methods like mdp.nodes.IdentityNode.__init__ have
the same __module__ attribute as the class to which they are
attached. This increases consistency, and has the additional advantage
that when epydoc looks at the methods and their __module__ attributes,
it is not confused by un-reachable modules.

Removes epydoc messages like:
Warning: __init__'s docstring is not a unicode string, but it contains
  non-ascii data -- treating it as latin-1.

ver. 2: add documentation string