nodes: Add ._new_covariance_matrix() for easy extension
It's awkward to write an extension tweaking extend a class'
.__init__() method, so we pull the CovarianceMatrix() call out into
its own method.  Tiziano wants extension initialization to effect
untrained (but already initialized) instances, so we postpone calling
the new method until we hit the training phase (with the hasattr()
check for the first training call).

Existing calls to CovarianceMatrix are not particularly consistent
with respect to their use of `bias`, which is in any case deprecated
in favor of ddof since NumPy 1.5.  I think this should be cleaned up
in MDP as well, but have left the initialization calls as they stand
during this commit's refactoring.