FIX: wrap sklearn algs in modules ending with "_"
For historical reasons we were not wrapping sklearn algs in modules with
names ending with "_". This restriction prevented us from wrapping algs
that we could have wrapped. We lift it now, since it does not seem to
have any bad side effects anymore.
A proper validation would be to really test that we can run the algs.