FIX: remove __future__ import in test_binode
The __future__ import is only needed in Python 3 < 3.5 (PEP 479).
It doesn't hurt with Python >= 3.5, but it breaks Python 2 compatibility.
To keep supporting Python 2 we can not use it. As far as I can tell, bimdp
was not dependent on the behaviour changed in PEP 479, so this change
should be safe.